Proof-of-Concept/requirements.txt
hwcounter==0.1.1
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.1
psutil==5.9.4
python_mbedtls==2.5.1
thespian==3.10.6